+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Changelog for NeoFS Node
 - `object search` CLI command is now the same as `object searchv2` (#3931)
 - Optimized integer comparison in SearchV2 (#3938)
 - Optimized number of nodes for SEARCH in EC containers (#3940)
+- Optimized post-initial placement replication (#3923)
 
 ### Removed
 - `policer.max_workers` configuration (#3920)